♻️ rename project projections
