+++ marketplace-domain/src/project/projections/lead_contributors.rs
